Moved table functions to utils and contraints to utils/constraints
